+++ frappe/database/utils.py
@@ -6,9 +6,9 @@ from functools import cached_property, wraps
 from types import NoneType
 
 import frappe
-from frappe.model.document import DocRef
 from frappe.query_builder.builder import MariaDB, Postgres
 from frappe.query_builder.functions import Function
+from frappe.types import DocRef
 
 if typing.TYPE_CHECKING:
 	from frappe.query_builder import DocType
