+++ frappe/model/base_document.py
@@ -350,8 +350,6 @@ class BaseDocument:
 	def get_valid_dict(
 		self, sanitize=True, convert_dates_to_str=False, ignore_nulls=False, ignore_virtual=False
 	) -> _dict:
-		from frappe.model.document import DocRef
-
 		d = _dict()
 		field_values = self.__dict__
