+++ frappe/types/__init__.py
@@ -1,0 +1,2 @@
+from .docref import DocRef
+from .frappedict import _dict
